fix: change the token-related methods directory
